Dedupe ST func_8018D8F0
