+++ src/st/cen/11280.c
@@ -352,14 +352,7 @@ s32 GetDistanceToPlayerY(void) {
     return yDistance;
 }
 
-s16 GetSideToPlayer(void) {
-    s16 var_a0 = g_CurrentEntity->posX.i.hi > PLAYER.posX.i.hi;
-
-    if (g_CurrentEntity->posY.i.hi > PLAYER.posY.i.hi) {
-        var_a0 |= 2;
-    }
-    return var_a0;
-}
+#include "../get_side_to_player.h"
 
 void MoveEntity() {
     g_CurrentEntity->posX.val += g_CurrentEntity->velocityX;
