+++ src/st/get_side_to_player.h
@@ -1,0 +1,14 @@
+/**
+ * Returns the player's side position relative to g_CurrentEntity
+ * 0 = Player is on the right side
+ * 1 = Player is on the left side
+ * 2 = Player is above
+ */
+s16 GetSideToPlayer(void) {
+    s16 var_a0 = g_CurrentEntity->posX.i.hi > PLAYER.posX.i.hi;
+
+    if (g_CurrentEntity->posY.i.hi > PLAYER.posY.i.hi) {
+        var_a0 |= 2;
+    }
+    return var_a0;
+}
